requirements.txt
tensorflow
numpy
pillow
opencv-python
ttkbootstrap
matplotlib
seaborn
pandas
